+++ apollo/src/main/graphql/com/hedvig/android/owldroid/graphql/profile.graphql
@@ -10,38 +10,6 @@ query Profile {
     ... CostFragment
     freeUntil
   }
-  contracts {
-    status {
-      ... ContractStatusFragment
-    }
-  }
-  balance {
-    failedCharges
-  }
-  chargeEstimation {
-    charge {
-      amount
-    }
-    discount {
-      amount
-    }
-    subscription {
-      amount
-    }
-  }
-  nextChargeDate
-  chargeHistory {
-    amount {
-      ...MonetaryAmountFragment
-    }
-    date
-  }
-  bankAccount {
-    ... BankAccountFragment
-  }
-  activePaymentMethods {
-    ... ActivePaymentMethodsFragment
-  }
   cashback {
     ... CashbackFragment
   }
@@ -49,23 +17,5 @@ query Profile {
     id
     name
     paragraph
-  }
-  redeemedCampaigns {
-    ... IncentiveFragment
-    owner {
-      displayName
-    }
-  }
-  referralInformation {
-    campaign {
-      code
-      incentive {
-        ... on MonthlyCostDeduction {
-          amount {
-            amount
-          }
-        }
-      }
-    }
   }
 }
