+++ testdata/src/main/java/com/hedvig/app/testdata/feature/profile/builders/ProfileDataBuilder.kt
@@ -13,7 +13,6 @@ data class ProfileDataBuilder(
     private val email: String? = "test@example.com",
     private val phoneNumber: String? = "07012345678",
     private val cost: CostFragment = CostBuilder().build(),
-    private val contracts: List<ProfileQuery.Contract> = emptyList()
 ) {
     fun build() = ProfileQuery.Data(
         member = ProfileQuery.Member(
@@ -29,25 +28,6 @@ data class ProfileDataBuilder(
                 costFragment = cost
             )
         ),
-        contracts = contracts,
-        balance = ProfileQuery.Balance(
-            failedCharges = null
-        ),
-        chargeEstimation = ProfileQuery.ChargeEstimation(
-            charge = ProfileQuery.Charge(
-                amount = "0.00"
-            ),
-            discount = ProfileQuery.Discount(
-                amount = "0.00"
-            ),
-            subscription = ProfileQuery.Subscription(
-                amount = "0.00"
-            )
-        ),
-        nextChargeDate = LocalDate.of(2020, 1, 1),
-        chargeHistory = emptyList(),
-        bankAccount = null,
-        activePaymentMethods = null,
         cashback = ProfileQuery.Cashback(
             fragments = ProfileQuery.Cashback.Fragments(
                 CashbackFragment(
@@ -58,12 +38,5 @@ data class ProfileDataBuilder(
             )
         ),
         cashbackOptions = emptyList(),
-        redeemedCampaigns = emptyList(),
-        referralInformation = ProfileQuery.ReferralInformation(
-            campaign = ProfileQuery.Campaign(
-                code = "TEST123",
-                incentive = null
-            )
-        )
-    )
+   )
 }
